// datalocker-ui/types.ts
export interface UseBalancesResponse {
  filBalance: bigint;
  usdfcBalance: bigint;
  filecoinWarmStorageBalance: bigint;
  filBalanceFormatted: number;
  usdfcBalanceFormatted: number;
  filecoinWarmStorageBalanceFormatted: number;
  persistenceDaysLeft: number;
  persistenceDaysLeftAtCurrentRate: number;
  isSufficient: boolean;
  isRateSufficient: boolean;
  isLockupSufficient: boolean;
  rateNeeded: bigint;
  totalLockupNeeded: bigint;
  depositNeeded: bigint;
  currentRateAllowanceGB: number;
  currentStorageGB: number;
  currentLockupAllowance: bigint;
}
// ...
export interface SectionProps {
  balances?: UseBalancesResponse;
  isLoading?: boolean;
}